Move locations schemas to configs.
