+++ src/ogd/common/configs/locations/DirectoryLocationSchema.py
@@ -2,7 +2,7 @@
 from pathlib import Path
 from typing import Any, Dict, Final, List, Optional, Self
 ## import local files
-from ogd.common.schemas.locations.LocationSchema import LocationSchema
+from ogd.common.configs.locations.LocationSchema import LocationSchema
 from ogd.common.utils.typing import JSONMap, Map
 
 ## @class DirectoryLocationSchema
