+++ src/ogd/common/configs/locations/LocationSchema.py
