+++ src/ogd/common/configs/locations/URLLocationSchema.py
@@ -2,7 +2,7 @@
 from urllib.parse import urlparse, ParseResult
 from typing import Any, Dict, Final, List, Optional, Self
 ## import local files
-from ogd.common.schemas.locations.LocationSchema import LocationSchema
+from ogd.common.configs.locations.LocationSchema import LocationSchema
 from ogd.common.utils.typing import JSONMap, Map
 
 ## @class URLLocationSchema
